requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.2.7
django-cors-headers==4.3.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.20
django-storages==1.14.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.2
easy-thumbnails==2.8.5
et-xmlfile==1.1.0
idna==3.4
lxml==4.9.3
numpy==1.26.1
oauthlib==3.2.2
odfpy==1.4.1
opencv-python==4.8.1.78
openpyxl==3.1.2
pandas==2.1.2
Pillow==10.1.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv
python3-openid==3.2.0
pytz==2023.3.post1
pyxlsb==1.0.10
requests==2.31.0
requests-oauthlib==1.3.1
setuptools==68.2.2
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
xlrd==2.0.1
